Attempt a reorganization
