+++ tests/game_spec.rb
