+++ tests/game_test.rb
